+++ src/screens/CartScreen/CartScreen.tsx
@@ -37,8 +37,10 @@ import PremiumCard from "../../components/PremiumCard";
 import { AccrodionPriceModel } from "../../models/AccrodionPriceModel";
 import { CheckBox } from "native-base";
 import { UserDataContext } from "../../provider/UserDataProvider";
-import ShipmentSection from "./ShipmentSection"
-import { Shipment } from "./Shipment"
+import ShipmentSection from "./ShipmentSection";
+import { Shipment } from "./Shipment";
+import { CartContext } from "../../context/cartStore";
+import { Types } from "../../context/cartReducer";
 
 type ShopScreenRouteProp = StackScreenProps<PurchaseStackParamList, "Cart">;
 
@@ -56,6 +58,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
   const isFocused = useIsFocused();
   const userDataStore = useContext(UserDataContext);
   const { userData } = userDataStore;
+  const { state, dispatch } = useContext(CartContext);
 
   useEffect(() => {
     getCart();
@@ -95,15 +98,24 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
   };
 
   const increaseProduct = async (itemId: string, quantity: number) => {
+    dispatch({
+      type: Types.Adjust,
+      payload: {
+        id: itemId,
+        quantity: quantity + 1,
+        shopId: route.params.shop.id,
+      },
+    });
     await CartDataSource.addToCartByShopId(
       route.params.shop.id,
       itemId,
       quantity + 1,
       payment,
       useSubsidize,
-      route.params.productBrand,
+      route.params.productBrand
     ).then((res: CartEntity) => {
       setCart(res);
+
       let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
         res.received_special_request_discounts
       );
@@ -122,7 +134,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
       quantity - 1,
       payment,
       useSubsidize,
-      route.params.productBrand,
+      route.params.productBrand
     ).then((res: CartEntity) => {
       setCart(res);
       let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
@@ -145,7 +157,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         quantity,
         payment,
         useSubsidize,
-        route.params.productBrand,
+        route.params.productBrand
       ).then((res: CartEntity) => {
         setCart(res);
         let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
@@ -163,12 +175,19 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
   };
 
   const removeItem = async (itemId: string) => {
+    dispatch({
+      type: Types.Delete,
+      payload: {
+        id: itemId,
+        shopId: route.params.shop.id,
+      },
+    });
     CartDataSource.removeItem(
       route.params.shop.id,
       itemId,
       payment,
       useSubsidize,
-      route.params.productBrand,
+      route.params.productBrand
     ).then((res: CartEntity) => {
       setCart(res);
       let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
@@ -183,7 +202,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
   };
 
   const handleShipmentChange = (s: Shipment) => {
-    setShipment(s)
+    setShipment(s);
   };
 
   const handlePayment = (payment: string) => {
@@ -193,7 +212,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         route.params.shop.id,
         payment,
         useSubsidize,
-        route.params.productBrand,
+        route.params.productBrand
       ).then((res: CartEntity) => {
         setCart(res);
         let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
@@ -211,7 +230,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         route.params.shop.id,
         payment,
         useSubsidize,
-        route.params.productBrand,
+        route.params.productBrand
       ).then((res: CartEntity) => {
         setCart(res);
         let discountSpecial: AccrodionPriceModel[] = formatAccrodion(
@@ -245,10 +264,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
     });
   };
 
-  const confirmOrder = (
-    shop: ShopEntity,
-    cart?: CartEntity
-  ) => {
+  const confirmOrder = (shop: ShopEntity, cart?: CartEntity) => {
     if (!cart?.selected_payment) {
       alert("กรุณาเลือกวิธีการชำระเงิน");
     } else if (!shipment) {
@@ -260,7 +276,7 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
         shipment,
         cart,
         cart.subsidize_discount,
-        route.params.productBrand,
+        route.params.productBrand
       ).then((res: OrderEntity) => {
         CartDataSource.clearCart(shop.id, route.params.productBrand);
         navigation.navigate("OrderSuccess", { data: res, cart });
@@ -618,12 +634,12 @@ const CartScreen: React.FC<ShopScreenRouteProp> = ({ navigation, route }) => {
                       {currencyFormat(cart.total_price, 2)}
                     </Text>
                   </View>
-                </View>     
+                </View>
                 <View style={styled.warpDelivery}>
-                  <ShipmentSection 
-                    shopId={route.params.shop.id} 
+                  <ShipmentSection
+                    shopId={route.params.shop.id}
                     onChange={(v) => {
-                      if(v) handleShipmentChange(v)
+                      if (v) handleShipmentChange(v);
                     }}
                     withDefault={true}
                   />